# httpapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: SemaDB API
  summary: No fuss vector database for AI
  description: >-
    This API provides a way to interact with a vector database, which is a type
    of database that stores data as high-dimensional vectors. Vectors are
    mathematical representations of features or attributes, and each vector has
    a certain number of dimensions, which can range from tens to thousands.

    Vector databases are used for a variety of tasks, such as similarity search,
    recommendation systems, and anomaly detection. They are particularly
    well-suited for tasks involving unstructured data, such as text, images, and
    audio.
  version: 0.1.0
  termsOfService: https://www.semafind.com/termsofservice
  contact:
    name: Semafind
    url: https://www.semafind.com/contact
    email: support@semafind.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
# ---------------------------
...
paths:
  /collections:
    summary: Endpoint for managing collections
    description: >-
      This endpoint covers operations for collections as a whole, such as
      creating a new collection or listing all collections.
    post:
      tags:
        - Collection
      summary: Create a new collection
      description: >-
        Creates a new collection if it does not already exist. The maximum
        number of collections per user is restricted based on the plan. Before
        you can insert and search points, you must create a collection.
      operationId: CreateCollection
      requestBody:
        description: The collection to create
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  $ref: '#/components/schemas/CollectionId'
                vectorSize:
                  $ref: '#/components/schemas/VectorSize'
                distanceMetric:
                  $ref: '#/components/schemas/DistanceMetric'
            examples:
              SampleCreateCollection:
                summary: Sample create collection
                description: A sample collection to create with id mycollection, vector size 2 and euclidean distance.
                value:
                  id: mycollection
                  vectorSize: 2
                  distanceMetric: euclidean
      responses:
        '200':
          $ref: '#/components/responses/SuccessfulMessageResponse'
          description: The collection was created
        '403':
          $ref: '#/components/responses/ErrorMessageResponse'
          description: The maximum collecton quota may be reached
        '409':
          $ref: '#/components/responses/ErrorMessageResponse'
          description: The collection already exists
    get:
      tags:
        - Collection
      summary: List user collections
      description: >-
        Returns a list of all collections for the current user. The list is not
        sorted by any value and the order may change between requests.
      operationId: ListCollection
      responses:
        '200':
          description: The list of collections
          content:
            application/json:
              schema:
                type: object
                properties:
                  collections:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          $ref: '#/components/schemas/CollectionId'
                        vectorSize:
                          $ref: '#/components/schemas/VectorSize'
                        distanceMetric:
                          $ref: '#/components/schemas/DistanceMetric'
              examples:
                SampleCollectionList:
                  summary: Sample collection list
                  description: A sample collection list with one collection
                  value:
                    collections:
                      - id: mycollection
                        vectorSize: 2
                        distanceMetric: euclidean
# ---------------------------
  /collections/{collectionId}:
    summary: Endpoint for managing a specific collection
    description: >-
      These endpoints interact with a single collection. The collection must
      exist before you can use these endpoints.
    parameters:
      - $ref: '#/components/parameters/CollectionId'
    get:
      tags:
        - Collection
      summary: Get the details of a collection
      description: >-
        This endpoint attempts to also list the shards currently available in
        the collection. Some shards may be temporarily unavailable. In that
        case, you can retry at a future time.
      operationId: GetCollection
      responses:
        '200':
          description: Collection with available shard details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    $ref: '#/components/schemas/CollectionId'
                  vectorSize:
                    $ref: '#/components/schemas/VectorSize'
                  distanceMetric:
                    $ref: '#/components/schemas/DistanceMetric'
                  shards:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        pointCount:
                          type: integer
              examples:
                SampleCollection:
                  summary: Sample collection
                  description: A sample collection with shard information
                  value:
                    id: mycollection
                    vectorSize: 2
                    distanceMetric: euclidean
                    shards:
                      - id: fff3a226-b9f8-4375-8dbd-1a240e000705
                        pointCount: 1000
        '503':
          $ref: '#/components/responses/ErrorMessageResponse'
          description: Some downstream components may be temporarily unavailable
    delete:
      tags:
        - Collection
      summary: Delete a collection
      description: >-
        Deletes a collection and all of its points. This operation is
        irreversible. If you want to delete only some points, use the bulk
        delete endpoint. If some shards are temporarily unavailable, the
        operation will still succeed, but some of the data will be deleted in
        the future.
      operationId: DeleteCollection
      responses:
        '200':
          $ref: '#/components/responses/SuccessfulMessageResponse'
          description: The collection and all of its data was deleted
        '202':
          $ref: '#/components/responses/SuccessfulMessageResponse'
          description: The collection was deleted, but some of the data will be deleted in the future
# ---------------------------
...
components:
  parameters:
    CollectionId:
      name: collectionId
      in: path
      description: The unique identifier of the collection
      required: true
      schema:
        $ref: '#/components/schemas/CollectionId'
  schemas:
    CollectionId:
      type: string
      title: Collection Id
      description: The unique identifier of the collection
      pattern: "^[a-z0-9]{3,16}$"
      examples:
        - mycollection
        - abc
    VectorSize:
      type: number
      description: The size of the vectors in the collection
      minimum: 1
      maximum: 2000
      examples:
        - 2
    DistanceMetric:
      type: string
      enum:
        - euclidean
        - cosine
...
  responses:
    SuccessfulMessageResponse:
      description: Operation was successful
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                description: A message indicating the result of the operation
            examples:
              - message: Operation successful
          examples:
            Success:
              summary: A successful response
              description: The operation was successful
              value:
                - message: success
    ErrorMessageResponse:
      description: Something unexpected happened
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
                description: An error message hopefully describing the problem
          examples:
            GeneralError:
              summary: General error
              description: Something unexpected happened
              value:
                - error: Something went wrong